Refactor utilities related to file in fileutil.py
